[NFC] Avoid violating same-filename rule in tests
…and modify resolveMagicNameConflicts() to diagnose any such violations instead of asserting.

Swift does not allow any two files in the same module to have the same filename, even if they are in different directories. However, this is enforced in the driver, so tests that invoke the frontend directly can violate it. Turns out that a couple of those snuck into the test suite at various points.

This commit updates those tests and causes the frontend to diagnose the duplicate filename error just as the driver would have.

NFC, since invoking the frontend directly is unsupported.