[android] Mark large_string_array as unsupported in Android-ARMv7
Seems that the test takes more than 5 minutes to finish, so the timeout
is triggered. Remove the test from the supported set for Android-ARMv7.
Instead of using XFAIL, use UNSUPPORTED to avoid spending 5 minutes
waiting for the timeout to occur. I don't think a lot of people will be
able to spend time to find out what's happening in a niche platform.

To make more clear what was happening, include a message if the
timeout.py script actually times out. Otherwise the error is a cryptic
"FileCheck error: '-' is empty.". Also fix the shebang and import the
print function from __future__ to avoid using the old print-chevron.