Work Around The Shared Module Cache's Lack of Concurrency Safety
The Clang Module cache is supposed to be concurrency safe by virtue of taking a number of file locks during PCM emission. In practice, the locking done by clang is insufficient. It appears that submodules can be mutated out from under the module loaders when multiple builds are in-flight. This causes spurious failures in downstream build jobs that can usually be solved by nuking the module cache and trying again - hoping against hope that your build system races in your favor.

When asked to emit PCHs, if it fails, give it one more try with a fresh module cache. This cache path is derived from the name of the module and the hash that is computed for the module cache for Swift Interface files for a little extra entropy. This ensures we have a truly fresh module cache.

The downsides to this approach are that we are trading compile time for stability. Rebuilding the module cache is an expensive proposition.

If Clang moves to explicit module builds, we can nuke this infrastructure.