[SIL] [Parser] Add new parers boilerplate visitors.
Adds basic infastructure and visitors that will be used to implement the
new parser.

Adds three classes: ReadSIL, CheckSIL, EmitSIL. These are
used to read information from a source file, check that the information
is valid, and parse the information into sil instructions respectively.

SILParser has been updated to use ReadSIL as the parser and to use all
three classes to parse instructions when possible.