Fix DSE and RLE unkown instruction handling.
DSE and RLE passes used to, in most cases, bail on all tracked locations
when they ran into an instruction that modified memory that they
couldn't identify. In most cases this memory address can be tracked to a
specific tracked location. If so, only that location has to be
invalidated.