Resolve master conflicts (#730)
* [build-script] Add the installation prefix to the toolchain path

In several places, it was assumed to be /usr and hard-coded to that. This makes sure
the installation prefix actually passed in is used.

* [stdlib] StaticString: extension per conformance

* Fix @available(macCatalyst N, iOS N+M) bug

Previously, availability checking computed a declaration’s availability as the intersection of all @available attributes with active platforms. This meant that Swift would not allow you to use an API that was available earlier in a child platform than in its parent until it was also available in the parent platform. That was incorrect.

This PR corrects availability checking to find the most specific @available attribute with an introduced version and use that instead.

Fixes rdar://60892534.

* [stdlib] Buildfix differentiation for OpenBSD.

New files were added in #30875 which did not include os(OpenBSD), so add
this.

add_swift_target_library in AddSwiftStdlib subsequently required
modification. _add_target_variant_link_flags likely needs modification as
well, but this is better suited to a separate PR.

* Handle multiple @available attributes correctly

The previous commit broke a promise made by the implementation it replaced: if there were two or more @available(introduced:) attributes for the same platform, the greatest version wins. This commit restores that property while still completely overriding the parent platform’s availability with a child platform’s.

* [Diagnostics] Provide anchors on demand instead of storing in `FailureDiagnostic`

This decouples `FailureDiagnostic` from expression.

* [Diagnostics] Switch `get{Raw}Anchor` to return `TypedNode`

In preparation to anchor `ConstraintLocator` from `TypedNode`
let's refactor diagnostics (which is the biggest user of locators)
to support `TypedNode` instead of `Expr *`.

* [Diagnostics] Switch `getConstraintLocator` variants to accept `TypedNode`

* [Diagnostics] Switch `getContextualType*` to use `TypedNode`

* [Diagnostics] Switch `getType` to use `TypedNode`

* [Diagnostics] Resolve const-ness problems introduced by `TypedNode`

Since `TypedNode` elements are all marked as `const` diagnostics
need to get some of the APIs adjusted to support passing `const Expr *`.

* Add compiler option to *disable* warnings-as-errors
`-no-warnings-as-errors`

This functionality is required for build systems to be able to overload/disable a given Swift project's preference of treating warnings as errors.

Resolves rdar://problem/35699776

* [Diagnostics] Add helper functions to work with anchors - `{castTo, getAs, is}Expr`

* [Diagnostics] Adjust `getCallInfo` to accept `TypedNode` instead of expression

* Avoid unnecessary performTypeChecking() in REPL

* [SIL Diagnostics] Improve diagnostics for capture of inout values
within escaping autoclosures.

<rdar://problem/60551844>

* [NFC] Refactor validateTBDIfNeeded

Inline the conditions for the astGuaranteedToCorrespondToSIL parameter and clean up the logic here to be more const-friendly. Also try to centralize and document the reasons we skip TBD validation.

* Revert "[Darwin] Further restrict inference of the simulator environment"

* [Gardening] Const-qualify CompilerInstance::getInvocation()

* [NFC] Formalize isModuleExternallyConsumed

Yank this parameter and document the historical scope creep of this value.

* More StringRef -> std::string conversion fixes.

This is needed for the master-next branch.

* Devirtualizer: fix a miscompile due to handling of cast instructions.

getInstanceWithExactDynamicType returns a new instance and for this the class decl has to be updated.

https://bugs.swift.org/browse/SR-12538
rdar://problem/61911112

* [NFC] Fix typo in SerializedModuleLoader.cpp

Co-authored-by: Butta <repo@butta.fastem.com>
Co-authored-by: Ben Rimmington <me@benrimmington.com>
Co-authored-by: Brent Royal-Gordon <brentdax@apple.com>
Co-authored-by: 3405691582 <dsk@google.com>
Co-authored-by: Pavel Yaskevich <pyaskevich@apple.com>
Co-authored-by: Artem Chikin <achikin@apple.com>
Co-authored-by: Rintaro Ishizaki <rishizaki@apple.com>
Co-authored-by: Ravi Kandhadai <rkandhadaimadhav@apple.com>
Co-authored-by: Robert Widmann <devteam.codafi@gmail.com>
Co-authored-by: Saleem Abdulrasool <compnerd@compnerd.org>
Co-authored-by: Mishal Shah <shahmishal@users.noreply.github.com>
Co-authored-by: Michael Forster <forster@google.com>
Co-authored-by: swift-ci <swift-ci@users.noreply.github.com>
Co-authored-by: Erik Eckstein <eeckstein@apple.com>
Co-authored-by: Ravi Kandhadai <37849400+ravikandhadai@users.noreply.github.com>
Co-authored-by: Pavel Yaskevich <xedin@apache.org>