Fix glibc.modulemap.gyb for WASI.
Two of the conditional blocks excluded for WASI contained unbalanced
braces; this can't be right.

I'm not sure how to test the WASI build, as glibc.modulemap currently
doesn't even get built on WASI (which is likely why this hasn't been
noticed). However, it does continue to build fine on Linux.