Extensible enumerations for non-resilient libraries.
Currently there is a feature mismatch between the resilient and
non-resilient dialects of the Swift language. In resilient libraries,
enumerations default to a mode usually called "open" or "extensible". In
this mode, calling code is required to add "@unknown default" clauses
when performing exhaustive switches over enumerations vended by the
library. This allows the library to add clauses to their enumerations
without breaking callers.

For non-resilient libraries, the default enumeration behaviour is to be
"frozen", or non-extensible. In this mode, callers of the library may
assume that no further case will be added to the enumeration without an
appropriate semantic versioning change, and so may exhaustively switch
over the enum without having to handle the possibility of a case they
did not see before.

This establishes two language dialects: one that defaults to frozen, and
one that defaults to extensible. Leaving aside whether this is a
positive state of affairs or not, however, these two dialects are not
equal.

This is because the resilient dialect may opt-in to the behaviour of the
non-resilient one by using the `@frozen` attribute. If a resilient
library applies this attribute to an enumeration they opt into a number
of behaviours, including exposing its size and direct access to the
enum's layout. However, they _also_ transition the exhaustive switching
behaviour: users are no longer required to add `@unknown default` to
their switch statements.

It is not currently possible for the non-resilient dialect to opt into
the extensible behaviour, however. No amount of work for a non-resilient
library can give them the ability to add cases to an enumeration without
incurring a major semantic version bump. There is no `@unfrozen`,
`@wet`, `@moist`, or other attribute that can be applied to an
enumeration to force this behaviour.

This makes the resilient language mode strictly _more_ powerful than the
non-resilient one. It also drastically neuters the power of enumerations
in non-resilient libraries. Common Swift patterns such as using
enumerations for errors, or to provide tagged union arguments to
methods, are not useful to non-resilient Swift libraries unless they are
extremely confident that the enumeration will never change. In essence,
the non-resilient language dialect forces developers to tag all
enumerations as `@frozen`. This is despite the guidance on `@frozen` in
the _resilient_ langauge dialect being to be extremely cautious and
judicious with its application.

To resolve the issue, this patch adds an attribute to enumerations that
allows them to opt-in to the extensible behaviour. This resolves the
feature incompatibility between the two language dialects, and makes it
easier to bring the two dialects together in future.