[TF] Use "@rpath" as install_name directory for standard library dylibs.
This sets "@rpath" as the directory in LC_DYLIB_ID for standard library dylibs.
This is important for toolchains that provide custom standard library modules or
that make changes to libswiftCore.dylib.

This patch should not be upstreamed to master, where Swift-in-the-OS is king and
it seems intended for Swift system libraries to be used.